Slight UI change to food facts page and bug fixed: properly gets all values from table and presents them in.

+++ NutritionTracker/FoodDisplay.cs
@@ -7,12 +7,12 @@ public class FoodDisplay
         Name = name;
     }
 
-    public FoodDisplay(string name, int value) : this(name)
+    public FoodDisplay(string name, float value) : this(name)
     {
         Value = value;
     }
 
     public string Name { get; set; }
-    public int Value { get; set; }
+    public float Value { get; set; }
     public Type TargeType { get; set; }
 }
