+++ NutritionTracker/NutritionFacts/FoodNutritionDetail.cs
@@ -9,12 +9,12 @@ class FoodNutritionDetail
 {
     public int food_code { get; set; }
     public string food_description { get; set; }
-    public int nutrient_value { get; set; }
+    public float nutrient_value { get; set; }
     public string nutrient_name { get; set; }
 
     public FoodNutritionDetail() { }
 
-    public FoodNutritionDetail(int food_code, string food_description, int nutrient_value, string nutrient_name) 
+    public FoodNutritionDetail(int food_code, string food_description, float nutrient_value, string nutrient_name) 
     {
         this.food_code = food_code;
         this.food_description = food_description;
